requirements.txt
python-telegram-bot
SQLAlchemy
mysql-connector-python
passlib
python-dotenv
apscheduler